stock-service/pyproject.toml
[tool.poetry]
name = "stock-service"
version = "0.1.0"
description = "Portfolio maker, stock analyzer and other services"
authors = ["Aleksa Mihajlovic <mihajlovic.aleksa@gmail.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.9.1"
python-dotenv = "^1.0.0"
asyncio = "^3.4.3"
pandas = "^2.2.0"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.41"
pydantic = "^2.11.5"
apscheduler = "^3.10.4"
ipykernel = "^6.30.1"
polars = "^1.32.3"
pyarrow = "^21.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"